refactor(cloudwatch-logs): Split index_recommender into 3 focused modules
Split the 1232-line index_recommender.py into three modules with clear
single responsibilities and zero circular dependencies:

- query_parser.py (274 lines): Regex patterns, CWLI/SQL/PPL parsing,
  language detection, field extraction, value stripping
- scoring.py (256 lines): Weights, constants, Pydantic result models,
  scoring engine, cardinality refinement
- index_recommender.py (463 lines): Analysis pipelines, API helpers,
  tool entry points

No behavioral changes. All 73 index recommender tests pass, full suite
of 444 tests pass.
